[buck2]REMOVE: redundant definition of orion tasks and targets in orion server
