+++ orion-server/src/model/orion_tasks.rs
@@ -1,42 +1,5 @@
-//! `SeaORM` Entity, @generated by sea-orm-codegen 1.1.19
-
-use api_model::buck2::{status::Status, types::ProjectRelativePath};
-use sea_orm::{IntoActiveModel, entity::prelude::*};
-use serde::{Deserialize, Serialize};
-
-#[derive(Clone, Debug, PartialEq, DeriveEntityModel, Eq, Serialize, Deserialize)]
-#[sea_orm(table_name = "orion_tasks")]
-pub struct Model {
-    #[sea_orm(primary_key, auto_increment = false)]
-    pub id: Uuid,
-    #[sea_orm(column_type = "JsonBinary")]
-    pub changes: Json,
-    pub repo_name: String,
-    pub cl: String,
-    pub created_at: DateTimeWithTimeZone,
-}
-
-#[derive(Copy, Clone, Debug, EnumIter, DeriveRelation)]
-pub enum Relation {
-    #[sea_orm(has_many = "super::build_events::Entity")]
-    BuildEvents,
-    #[sea_orm(has_many = "super::build_targets::Entity")]
-    BuildTargets,
-}
-
-impl Related<super::build_events::Entity> for Entity {
-    fn to() -> RelationDef {
-        Relation::BuildEvents.def()
-    }
-}
-
-impl Related<super::build_targets::Entity> for Entity {
-    fn to() -> RelationDef {
-        Relation::BuildTargets.def()
-    }
-}
-
-impl ActiveModelBehavior for ActiveModel {}
+use serde_json::to_value;
+use serde_json::Value;
 
 pub struct OrionTask;
 
@@ -66,5 +29,26 @@ impl OrionTask {
         let task_model = Self::create_task(task_id, cl_link, repo, changes)
             .map_err(|e| DbErr::Custom(e.to_string()))?;
         task_model.into_active_model().insert(db).await
+    }
+}
+
+#[derive(ToSchema, Serialize)]
+pub struct OrionTaskDTO {
+    pub id: String,
+    pub changes: Value,
+    pub repo_name: String,
+    pub cl: String,
+    pub created_at: String,
+}
+
+impl From<&callisto::orion_tasks::Model> for OrionTaskDTO {
+    fn from(model: &callisto::orion_tasks::Model) -> Self {
+        Self {
+            id: model.id.to_string(),
+            changes: model.changes.clone(),
+            repo_name: model.repo_name.clone(),
+            cl: model.cl.clone(),
+            created_at: model.created_at.with_timezone(&Utc).to_string(),
+        }
     }
 }
